+++ main/AppInfo.cs
@@ -2,8 +2,8 @@
 
 public static class AppInfo
 {
-    // public const string Version = "2026.21.6";
-    public const string Version = "2026.21.6";
+    // public const string Version = "2026.22.0";
+    public const string Version = "2026.22.0";
     public const string ContactEmail = "vrcn@shinyflvres.com";
     public const string Website = "vrcn.shinyflvres.com";
     public const string UserAgent = $"VRCNext/{Version} ({ContactEmail})";